+++ source/dotnet/Measurements.WebApi/Controllers/MeasurementsController.cs
@@ -11,15 +11,15 @@ namespace Energinet.DataHub.Measurements.WebApi.Controllers;
 
 [ApiController]
 [Authorize]
-[ApiVersion(2.0, Deprecated = true)]
-[ApiVersion(3.0)]
+[ApiVersion(3.0, Deprecated = true)]
+[ApiVersion(4.0)]
 [Route("v{v:apiVersion}/measurements")]
 public class MeasurementsController(
     IMeasurementsHandler measurementsHandler, ILogger<MeasurementsController> logger, IJsonSerializer jsonSerializer)
     : ControllerBase
 {
-    [MapToApiVersion(2.0)]
     [MapToApiVersion(3.0)]
+    [MapToApiVersion(4.0)]
     [HttpGet("forPeriod")]
     public async Task<IActionResult> GetByPeriodAsync([FromQuery] GetByPeriodRequest request)
     {
@@ -43,19 +43,21 @@ public class MeasurementsController(
     }
 
     [MapToApiVersion(3.0)]
+    [MapToApiVersion(4.0)]
     [HttpGet("currentForPeriod")]
     public Task<IActionResult> GetCurrentByPeriodAsync([FromQuery] GetByPeriodRequest request)
     {
         return Task.FromResult<IActionResult>(Accepted("This endpoint is not implemented yet."));
     }
 
-    [MapToApiVersion(2.0)]
-    [HttpGet("aggregatedByMonth")]
-    public async Task<IActionResult> GetAggregatedByDateAsyncV2([FromQuery] GetAggregatedByDateRequest request)
+    [MapToApiVersion(3.0)]
+    [HttpGet("aggregatedByDate")]
+    [Obsolete("v4.0 is deprecated. Use v4.0 instead.")]
+    public async Task<IActionResult> GetAggregatedByDateAsyncV3([FromQuery] GetAggregatedByDateRequest request)
     {
         try
         {
-            var aggregatedByMonth = await measurementsHandler.GetAggregatedByDateAsync(request);
+            var aggregatedByMonth = await measurementsHandler.GetAggregatedByDateAsyncV3(request);
             var result = jsonSerializer.Serialize(aggregatedByMonth);
 
             return Ok(result);
@@ -72,29 +74,7 @@ public class MeasurementsController(
         }
     }
 
-    [MapToApiVersion(2.0)]
-    [HttpGet("aggregatedByYear")]
-    public async Task<IActionResult> GetAggregatedByMonthAsyncV2([FromQuery] GetAggregatedByMonthRequest request)
-    {
-        try
-        {
-            var aggregatedByYear = await measurementsHandler.GetAggregatedByMonthAsync(request);
-            var result = jsonSerializer.Serialize(aggregatedByYear);
-
-            return Ok(result);
-        }
-        catch (MeasurementsNotFoundException e)
-        {
-            logger.LogInformation(
-                "Aggregation by year and month not found for metering point id {MeteringPointId} during {Year}",
-                request.MeteringPointId.Sanitize(),
-                request.Year);
-
-            return NotFound(e.Message);
-        }
-    }
-
-    [MapToApiVersion(3.0)]
+    [MapToApiVersion(4.0)]
     [HttpGet("aggregatedByDate")]
     public async Task<IActionResult> GetAggregatedByDateAsync([FromQuery] GetAggregatedByDateRequest request)
     {
@@ -118,6 +98,7 @@ public class MeasurementsController(
     }
 
     [MapToApiVersion(3.0)]
+    [MapToApiVersion(4.0)]
     [HttpGet("aggregatedByMonth")]
     public async Task<IActionResult> GetAggregatedByMonthAsync([FromQuery] GetAggregatedByMonthRequest request)
     {
@@ -140,6 +121,7 @@ public class MeasurementsController(
     }
 
     [MapToApiVersion(3.0)]
+    [MapToApiVersion(4.0)]
     [HttpGet("aggregatedByYear")]
     public async Task<IActionResult> GetAggregatedByYearAsync([FromQuery] GetAggregatedByYearRequest request)
     {
@@ -161,6 +143,7 @@ public class MeasurementsController(
     }
 
     [MapToApiVersion(3.0)]
+    [MapToApiVersion(4.0)]
     [HttpGet("aggregatedByPeriod")]
     public async Task<IActionResult> GetAggregatedByPeriodAsync([FromQuery] GetAggregatedByPeriodRequest request)
     {